@@ -10,21 +10,20 @@
     }
   ],
   "require": {
-    "php": ">=7.3",
+    "php": ">=8.0",
     "php-amqplib/php-amqplib": "^2.8",
     "aws/aws-sdk-php": "^3.69",
     "ext-json": "*"
   },
   "require-dev": {
-    "phpunit/phpunit": "^8.0",
-    "mockery/mockery": "^1.2",
-    "jakub-onderka/php-parallel-lint": "^1.0",
-    "phpro/grumphp": "^0.15",
-    "sensiolabs/security-checker": "^5.0",
     "nikic/php-parser": "^4.0",
-    "povils/phpmnd": "^2.0",
     "squizlabs/php_codesniffer": "^3.4",
-    "doctrine/coding-standard": "^6.0"
+    "phpunit/phpunit": "^9.5",
+    "phpro/grumphp": "^1.13",
+    "doctrine/coding-standard": "^10.0",
+    "mockery/mockery": "^1.5",
+    "povils/phpmnd": "^3.0",
+    "php-parallel-lint/php-parallel-lint": "^1.3"
   },
   "autoload": {
     "psr-4": {
@@ -40,5 +39,11 @@
   "suggest": {
     "ext-sockets": "Remove socket warnings when using RabbitMQ"
   },
-  "prefer-stable": true
+  "prefer-stable": true,
+  "config": {
+    "allow-plugins": {
+      "dealerdirect/phpcodesniffer-composer-installer": true,
+      "phpro/grumphp": true
+    }
+  }
 }
