@@ -11,42 +11,35 @@ use SykesCottages\Qu\Message\Contract\MessageHandler;
 
 final class Queue implements Consumable, MessageHandler
 {
-    /** @var string */
-    private $name;
+    private Queue $queue;
 
-    /** @var Queue */
-    private $queue;
-
-    public function __construct(string $name, QueueInterface $queue)
+    public function __construct(private string $name, QueueInterface $queue)
     {
-        $this->name  = $name;
         $this->queue = $queue;
     }
 
-    public function getQueueName() : string
+    public function getQueueName(): string
     {
         return $this->name;
     }
 
-    /**
-     * @param string[] $body
-     */
-    public function queueMessage(array $body) : void
+    /** @param string[] $body */
+    public function queueMessage(array $body): void
     {
         $this->queue->queueMessage($this->name, $body);
     }
 
-    public function consume(callable $callback, callable $idleCallback) : void
+    public function consume(callable $callback, callable $idleCallback): void
     {
         $this->queue->consume($this->name, $callback, $idleCallback);
     }
 
-    public function acknowledge(Message $message) : void
+    public function acknowledge(Message $message): void
     {
         $this->queue->acknowledge($this->name, $message);
     }
 
-    public function reject(Message $message, string $errorMessage = '') : void
+    public function reject(Message $message, string $errorMessage = ''): void
     {
         $this->queue->reject($this->name, $message, $errorMessage);
     }
