@@ -10,35 +10,30 @@ use SykesCottages\Qu\Message\Contract\Message;
 
 abstract class Consumer
 {
-    /** @var bool */
-    protected $exitRequested;
+    protected bool $exitRequested;
 
-    /** @var Queue */
-    protected $queue;
-
-    public function __construct(Queue $queue)
+    public function __construct(protected Queue $queue)
     {
-        $this->queue = $queue;
     }
 
     /**
      * Implement this function to process the message
      */
-    abstract public function process(Message $message) : void;
+    abstract public function process(Message $message): void;
 
-    public function start() : void
+    public function start(): void
     {
         $this->queue->consume([$this, 'process'], [$this, 'idle']);
     }
 
-    public function idle() : void
+    public function idle(): void
     {
         if ($this->exitRequested) {
             throw new ExitRequested($this->queue->getQueueName());
         }
     }
 
-    public function requestExit() : void
+    public function requestExit(): void
     {
         $this->exitRequested = true;
     }
