@@ -10,6 +10,7 @@ use SykesCottages\Qu\Message\Contract\Message;
 use SykesCottages\Qu\Message\SQSMessage;
 use Tests\Functional\Connector\Stubs\SQSCallable;
 use Tests\Functional\FunctionalTestCase;
+
 use function current;
 use function getenv;
 
@@ -23,12 +24,11 @@ class SQSTest extends FunctionalTestCase
 
     private const QUEUE_NAME = 'test';
 
-    /** @var SQS */
-    private $sqs;
-    /** @var string */
-    private $testingQueueUrl;
+    private SQS $sqs;
 
-    public function setUp() : void
+    private string $testingQueueUrl;
+
+    public function setUp(): void
     {
         $this->sqs = new SQS([
             'service' => 'sqs',
@@ -55,7 +55,7 @@ class SQSTest extends FunctionalTestCase
         ]);
     }
 
-    public function testWeCanConnectToSQSAndReturnAListOfQueueUrls() : void
+    public function testWeCanConnectToSQSAndReturnAListOfQueueUrls(): void
     {
         $activeQueues = $this->sqs->listQueues();
 
@@ -64,7 +64,7 @@ class SQSTest extends FunctionalTestCase
         $this->assertCount(self::DEFAULT_NUMBER_OF_URLS, $urls);
     }
 
-    public function testWeCanAcknowledgeAMessageInTheQueue() : void
+    public function testWeCanAcknowledgeAMessageInTheQueue(): void
     {
         $this->addMessageToQueue();
 
@@ -77,7 +77,7 @@ class SQSTest extends FunctionalTestCase
         $this->assertEmpty($this->getMessages($this->testingQueueUrl));
     }
 
-    public function testWeCanRejectAMessageInTheQueue() : void
+    public function testWeCanRejectAMessageInTheQueue(): void
     {
         $this->addMessageToQueue();
 
@@ -94,39 +94,37 @@ class SQSTest extends FunctionalTestCase
         $this->assertCount(1, $messages);
     }
 
-    public function testWeCanCallTheCallbackFunctionOnConsume() : void
+    public function testWeCanCallTheCallbackFunctionOnConsume(): void
     {
         $this->addMessageToQueue();
 
         $this->sqs->consume(
             $this->testingQueueUrl,
-            function (Message $message) : void {
+            function (Message $message): void {
                 $this->assertFunctionHasBeenCalled();
                 $this->assertInstanceOf(SQSMessage::class, $message);
             },
-            function () : void {
+            function (): void {
                 $this->assertFunctionIsNotCalled();
-            }
+            },
         );
     }
 
-    public function testWeCanCallTheIdleCallbackFunctionOnConsume() : void
+    public function testWeCanCallTheIdleCallbackFunctionOnConsume(): void
     {
         $this->sqs->consume(
             $this->testingQueueUrl,
-            function (Message $message) : void {
+            function (Message $message): void {
                 $this->assertFunctionIsNotCalled();
             },
-            function () : void {
+            function (): void {
                 $this->assertFunctionHasBeenCalled();
-            }
+            },
         );
     }
 
-    /**
-     * @return int[][]
-     */
-    public function retrieveMinAndMaxDataProvider() : array
+    /** @return int[][] */
+    public function retrieveMinAndMaxDataProvider(): array
     {
         return [
             'Test if we specify more than the MAX it will cap to the MAX value allowed by SQS' => [
@@ -140,13 +138,11 @@ class SQSTest extends FunctionalTestCase
         ];
     }
 
-    /**
-     * @dataProvider retrieveMinAndMaxDataProvider
-     */
+    /** @dataProvider retrieveMinAndMaxDataProvider */
     public function testWeCanOnlyRetrieveMessagesBetweenTheMaxAndMin(
         int $numberOfMessagesToConsumeAtATime,
-        int $expectedNumberOfCallbackCalls
-    ) : void {
+        int $expectedNumberOfCallbackCalls,
+    ): void {
         $this->addMultipleMessagesToQueue(100);
 
         $this->sqs->setQueueOptions(['maxNumberOfMessagesPerConsume' => $numberOfMessagesToConsumeAtATime]);
@@ -159,13 +155,13 @@ class SQSTest extends FunctionalTestCase
         $this->sqs->consume(
             $this->testingQueueUrl,
             $mock,
-            function () : void {
+            function (): void {
                 $this->assertFunctionIsNotCalled();
-            }
+            },
         );
     }
 
-    private function addMultipleMessagesToQueue(int $messagesToSend) : void
+    private function addMultipleMessagesToQueue(int $messagesToSend): void
     {
         $countOfMessages = 0;
         while ($countOfMessages < $messagesToSend) {
@@ -174,15 +170,13 @@ class SQSTest extends FunctionalTestCase
         }
     }
 
-    private function addMessageToQueue() : void
+    private function addMessageToQueue(): void
     {
         $this->sqs->queueMessage($this->testingQueueUrl, ['example' => 'test']);
     }
 
-    /**
-     * @return SQSMessage[]
-     */
-    private function getMessages(string $queueUrl) : array
+    /** @return SQSMessage[] */
+    private function getMessages(string $queueUrl): array
     {
         $message = $this->sqs->receiveMessage([
             'QueueUrl' => $queueUrl,
