@@ -12,18 +12,16 @@ use Tests\Unit\UnitTestCase;
 
 class RabbitMQMessageTest extends UnitTestCase
 {
-    /** @var AMQPMessage */
-    private $amqpMessage;
-    /** @var RabbitMQMessage */
-    private $rabbitMqMessage;
+    private AMQPMessage $amqpMessage;
+    private RabbitMQMessage $rabbitMqMessage;
 
-    public function setUp() : void
+    public function setUp(): void
     {
         $this->amqpMessage     = Mockery::mock(AMQPMessage::class);
         $this->rabbitMqMessage = new RabbitMQMessage($this->amqpMessage);
     }
 
-    public function testTheBodyIsReturnedAsAnAssociativeArray() : void
+    public function testTheBodyIsReturnedAsAnAssociativeArray(): void
     {
         $this->amqpMessage
             ->shouldReceive('getBody')
@@ -36,7 +34,7 @@ class RabbitMQMessageTest extends UnitTestCase
         $this->assertSame($expectedResult, $this->rabbitMqMessage->getBody());
     }
 
-    public function testTheBodyIsEmptyIfBlankJSONStringIsReturned() : void
+    public function testTheBodyIsEmptyIfBlankJSONStringIsReturned(): void
     {
         $this->amqpMessage
             ->shouldReceive('getBody')
@@ -47,7 +45,7 @@ class RabbitMQMessageTest extends UnitTestCase
         $this->assertEmpty($this->rabbitMqMessage->getBody());
     }
 
-    public function testGetDeliveryInfoChannelReturnsWithAChannel() : void
+    public function testGetDeliveryInfoChannelReturnsWithAChannel(): void
     {
         $this->amqpMessage
             ->shouldReceive('get')
@@ -58,7 +56,7 @@ class RabbitMQMessageTest extends UnitTestCase
         $this->assertInstanceOf(AMQPChannel::class, $this->rabbitMqMessage->getDeliveryInfoChannel());
     }
 
-    public function testGetDeliveryTagReturnsWithString() : void
+    public function testGetDeliveryTagReturnsWithString(): void
     {
         $randomDeliveryTag = 'THIS IS A RANDOM STRING';
 
