@@ -11,12 +11,13 @@ class SQSMessageTest extends UnitTestCase
 {
     private const RECEIPT_HANDLE = 'this-is-the-receipt-handle';
 
-    /** @var string[] */
-    private $message;
-    /** @var SQSMessage */
-    private $sqsMessage;
+    //@codingStandardsIgnoreStart
+    private array $message;
+    //@codingStandardsIgnoreEnd
 
-    public function setUp() : void
+    private SQSMessage $sqsMessage;
+
+    public function setUp(): void
     {
         $this->message = [
             'Body' => '{"test": "example"}',
@@ -26,19 +27,19 @@ class SQSMessageTest extends UnitTestCase
         $this->sqsMessage = new SQSMessage($this->message);
     }
 
-    public function testTheBodyIsReturnedAsAnAssociativeArray() : void
+    public function testTheBodyIsReturnedAsAnAssociativeArray(): void
     {
         $expectedResult = ['test' => 'example'];
 
         $this->assertSame($expectedResult, $this->sqsMessage->getBody());
     }
 
-    public function testGetReceiptHandleReturnsTheCorrectString() : void
+    public function testGetReceiptHandleReturnsTheCorrectString(): void
     {
         $this->assertSame(self::RECEIPT_HANDLE, $this->sqsMessage->getReceiptHandle());
     }
 
-    public function testRawMessageReturnsTheEntireMessage() : void
+    public function testRawMessageReturnsTheEntireMessage(): void
     {
         $this->assertSame($this->message, $this->sqsMessage->getRawMessage());
     }
