@@ -6,8 +6,8 @@ namespace SykesCottages\Qu\Message\Contract;
 
 interface Message
 {
-    /**
-     * @return string[]|null
-     */
-    public function getBody() : ?array;
+    // @codingStandardsIgnoreStart
+    /** @return mixed array */
+    public function getBody(): array;
+    // @codingStandardsIgnoreEnd
 }
