@@ -6,7 +6,7 @@ namespace SykesCottages\Qu\Message\Contract;
 
 interface MessageHandler
 {
-    public function acknowledge(Message $message) : void;
+    public function acknowledge(Message $message): void;
 
-    public function reject(Message $message, string $errorMessage = '') : void;
+    public function reject(Message $message, string $errorMessage = ''): void;
 }
