@@ -8,6 +8,7 @@ use SykesCottages\Qu\Connector\RabbitMQ;
 use SykesCottages\Qu\Message\Contract\Message;
 use SykesCottages\Qu\Message\RabbitMQMessage;
 use Tests\Functional\RabbitMQTestCase;
+
 use function getenv;
 
 class RabbitMQTest extends RabbitMQTestCase
@@ -16,26 +17,24 @@ class RabbitMQTest extends RabbitMQTestCase
 
     private const DEAD_LETTER_QUEUE_NAME = 'dead_letter';
 
-    public function setUp() : void
+    public function setUp(): void
     {
         $this->rabbitMq = new RabbitMQ(
             getenv('RABBIT_MQ_HOST'),
             (int) getenv('RABBIT_MQ_PORT'),
             getenv('RABBIT_MQ_USER'),
-            getenv('RABBIT_MQ_PASSWORD')
+            getenv('RABBIT_MQ_PASSWORD'),
         );
 
         $this->channel = $this->rabbitMq->channel();
     }
 
-    public function testWeCanConnectToTheRabbitMQServer() : void
+    public function testWeCanConnectToTheRabbitMQServer(): void
     {
-        $this->assertTrue(
-            $this->rabbitMq->isConnected()
-        );
+        $this->assertTrue($this->rabbitMq->isConnected());
     }
 
-    public function testWeCanAcknowledgeMessageAndDeleteItFromTheQueue() : void
+    public function testWeCanAcknowledgeMessageAndDeleteItFromTheQueue(): void
     {
         $this->addMessageToQueue();
 
@@ -44,7 +43,7 @@ class RabbitMQTest extends RabbitMQTestCase
         $this->assertQueueIsEmpty(self::QUEUE_NAME);
     }
 
-    public function testWeCanRejectAMessageAndSendItToTheDeadLetterQueue() : void
+    public function testWeCanRejectAMessageAndSendItToTheDeadLetterQueue(): void
     {
         $this->addMessageToQueue();
 
@@ -54,7 +53,7 @@ class RabbitMQTest extends RabbitMQTestCase
         $this->assertQueueHasAMessage(self::DEAD_LETTER_QUEUE_NAME);
     }
 
-    public function testWeCanCallTheCallbackFunctionWhenWeHaveAMessage() : void
+    public function testWeCanCallTheCallbackFunctionWhenWeHaveAMessage(): void
     {
         $this->rabbitMq->setQueueOptions([
             'blockingConsumer' => false,
@@ -65,17 +64,17 @@ class RabbitMQTest extends RabbitMQTestCase
 
         $this->rabbitMq->consume(
             self::QUEUE_NAME,
-            function (Message $message) : void {
+            function (Message $message): void {
                 $this->assertFunctionHasBeenCalled();
                 $this->assertInstanceOf(RabbitMQMessage::class, $message);
             },
-            function () : void {
+            function (): void {
                 $this->assertFunctionIsNotCalled();
-            }
+            },
         );
     }
 
-    private function addMessageToQueue() : void
+    private function addMessageToQueue(): void
     {
         $this->rabbitMq->queueMessage(self::QUEUE_NAME, ['example' => 'test']);
     }
