@@ -10,30 +10,24 @@ use Tests\Unit\UnitTestCase;
 
 class InvalidMessageTypeExceptionTest extends UnitTestCase
 {
-    public function testExceptionMatchesLogicExceptionClass() : void
+    public function testExceptionMatchesLogicExceptionClass(): void
     {
-        $exception = new InvalidMessageType(
-            self::class
-        );
+        $exception = new InvalidMessageType(self::class);
 
         $this->assertInstanceOf(LogicException::class, $exception);
     }
 
-    /**
-     * @dataProvider classMessageDataProvider
-     */
+    /** @dataProvider classMessageDataProvider */
     public function testExceptionProducesCorrectMessageBasedOnClassName(
         string $className,
-        string $expectedResult
-    ) : void {
+        string $expectedResult,
+    ): void {
         $exception = new InvalidMessageType($className);
         $this->assertSame($expectedResult, $exception->getMessage());
     }
 
-    /**
-     * @return string[][]
-     */
-    public function classMessageDataProvider() : array
+    /** @return string[][] */
+    public function classMessageDataProvider(): array
     {
         return [
             'Test when class name is passed the full path is returned in the error message' => [
